CDAP-7021 fix race and optimize first cache attempt
Fixed a race condition with checking whether the cache exists.
Also added an optimization if many programs of the same type launch
at the same time before the cache is populated. It will now only
launch the first program of that type and have the others wait for
the cache to be populated. This is an optimization in case there
are hundreds of workflows that are all scheduled to start at the
same time.